do not allow deletion of domain is it's associated with aws/gcp/azure
