+++ servers/zms/conf/zms.properties
@@ -552,3 +552,10 @@ athenz.zms.no_auth_uri_list=/zms/v1/schema
 # have the same concerns as with updated objects, so we're allowing a much
 # larger offset.
 #athenz.zms.review_date_offset_days_new_objects=365
+
+# The setting specifies comma separated list of domain meta attributes that
+# must be empty when a domain is deleted. Supported values for now are the
+# meta attributes for the public cloud providers - aws/gcp/azure settings
+# (e.g. account,gcpproject,azuresubscription). The server will validate that
+# the domain does not have any of these attributes set when the domain is deleted.
+#athenz.zms.domain_delete_meta_attributes=
